requirements.txt
pyaudio==0.2.14
azure-cognitiveservices-speech==1.37.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0
python-docx==0.8.11
